Fix compiler warnings, cut redundancy
